pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name            = "consensusflow"
version         = "0.1.0"
description     = "Multi-model verification pipeline — Proposer → Auditor → Resolver"
readme          = "README.md"
license         = { text = "MIT" }
requires-python = ">=3.9"
keywords        = ["llm", "verification", "hallucination", "consensus", "litellm", "ai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Framework :: AsyncIO",
]

dependencies = [
    "litellm>=1.40.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
consensusflow = "consensusflow.cli:main"

[project.urls]
Homepage      = "https://omer.github.io/consensusflow"
Repository    = "https://github.com/omer/consensusflow"
Documentation = "https://github.com/omer/consensusflow#readme"
"Bug Tracker" = "https://github.com/omer/consensusflow/issues"

[tool.setuptools.packages.find]
where   = ["."]
include = ["consensusflow*"]

[tool.setuptools.package-data]
"consensusflow.prompts" = ["*.md", "*.txt"]

# ── Ruff (linter) ─────────────────────────────────────────────
[tool.ruff]
line-length    = 99
target-version = "py39"

[tool.ruff.lint]
select  = ["E", "F", "I", "UP", "B", "SIM"]
ignore  = ["E501"]

# ── Pytest ─────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths    = ["tests"]

# ── Mypy ───────────────────────────────────────────────────────
[tool.mypy]
python_version         = "3.9"
strict                 = false
ignore_missing_imports = true
